+++ dist_node/inline/state.js
@@ -1,23 +1,27 @@
 /*
- * THIS FILE IS AUTO GENERATED FROM 'lib/inline/state.kep'
+ * THIS FILE IS AUTO GENERATED from 'lib/inline/state.kep'
  * DO NOT EDIT
-*/
-"use strict";
+*/"use strict";
 var record = require("bes")["record"],
     hamt = require("hamt"),
-    __o = require("./binding"),
-    Binding = __o["Binding"],
-    MUTABLE = __o["MUTABLE"],
-    IMMUTABLE = __o["IMMUTABLE"],
-    State, getBinding, addBinding, addWorking, push, pop, getGlobals, addGlobal;
-(State = record.declare(null, ["bindings", "working", "globals", "outer"]));
-(State.empty = new(State)(hamt.empty, hamt.empty, hamt.empty, null));
+    __o = require("../fun"),
+    concat = __o["concat"],
+    __o0 = require("./binding"),
+    Binding = __o0["Binding"],
+    MUTABLE = __o0["MUTABLE"],
+    IMMUTABLE = __o0["IMMUTABLE"],
+    State, getBinding, addBinding, addWorking, push, pop, getGlobals, addGlobal, addLocals, getLocals, pushLocals,
+        popLocals;
+(State = record.declare(null, ["bindings", "working", "globals", "outer", "locals"]));
+(State.empty = new(State)(hamt.empty, hamt.empty, hamt.empty, null, [
+    [], null
+]));
 (push = (function(s) {
     return s.setOuter(s)
         .setWorking(hamt.empty);
 }));
-var nullWorking = hamt.fold.bind(null, (function(p, __o0) {
-    var key = __o0["key"];
+var nullWorking = hamt.fold.bind(null, (function(p, __o1) {
+    var key = __o1["key"];
     return hamt.set(key, null, p);
 }));
 (pop = (function(s) {
@@ -43,6 +47,20 @@ var y = hamt.keys;
 (addGlobal = (function(name, s) {
     return s.setGlobals(hamt.set(name, name, s.globals));
 }));
+(addLocals = (function(locals, s) {
+    return s.setLocals([concat(locals, s.locals[0]), s.locals[1]]);
+}));
+(getLocals = (function(s) {
+    return s.locals[0];
+}));
+(pushLocals = (function(s) {
+    return s.setLocals([
+        [], s.locals
+    ]);
+}));
+(popLocals = (function(s) {
+    return s.setLocals(s.locals[1]);
+}));
 (exports["State"] = State);
 (exports["getBinding"] = getBinding);
 (exports["addBinding"] = addBinding);
@@ -51,3 +69,7 @@ var y = hamt.keys;
 (exports["pop"] = pop);
 (exports["getGlobals"] = getGlobals);
 (exports["addGlobal"] = addGlobal);
+(exports["addLocals"] = addLocals);
+(exports["getLocals"] = getLocals);
+(exports["pushLocals"] = pushLocals);
+(exports["popLocals"] = popLocals);