+++ dist_node/stages/lexical.js
@@ -1,19 +1,20 @@
 /*
- * THIS FILE IS AUTO GENERATED FROM 'lib/stages/lexical.kep'
+ * THIS FILE IS AUTO GENERATED from 'lib/stages/lexical.kep'
  * DO NOT EDIT
-*/
-"use strict";
+*/"use strict";
 var Error = require("akh")["error"],
+    __o = require("bes")["object"],
+    setProperty = __o["setProperty"],
     lexical = require("../lexical/lexical"),
     check, builtins = ["Array", "Boolean", "Date", "decodeURI", "decodeURIComponent", "encodeURI", "encodeURIComponent",
         "Error", "eval", "EvalError", "Function", "Infinity", "isFinite", "isNaN", "JSON", "Math", "NaN", "Number",
         "Object", "parseInt", "parseFloat", "RangeError", "ReferenceError", "RegExp", "String", "SyntaxError",
         "TypeError", "undefined", "URIError"
     ];
-(check = (function(__o) {
-    var options = __o["options"],
-        tree = __o["tree"],
-        data = __o["data"];
+(check = (function(input) {
+    var options = input["options"],
+        tree = input["tree"],
+        data = input["data"];
     return lexical.check(tree, ((options && options.globals) || builtins), data)
         .map((function(__o0) {
             var tree0 = __o0["tree"],